Fix shared volume in docker-compose template for /awx_devel
So, the key here was setting the selinux flag.
It might be able to be set to Z instead of 'z', but I'll need to test.
'Z' = share with just this container
'z' = share across other containers

Fix volume mounts in docker-compose and fallout from it

The relative paths in the volume mounts in the docker-compose
file/template were strange. I'm not sure if they were intentional,
but the relative paths are supposed to be from where the docker-
compose yml file lives. This file is generated and might end up
in a couple of different places, so I'm not sure if the relative
paths were always right. At any rate, I started fixing them so
that they were on-spec: relative bind/volume mounts aren't exactly
supported, but when they are used they need to always start with
`./` to avoid them looking like named volume mounts :/.

So, as I was doing that, some other stuff came to light, and I've
been trying to simplify and fix that stuff in this commit as well
as continuing to support and fix the settable sources (what was
previously hard-coded as _sources) and get things working with
rootless podman.